added a single file script to convert .res files to .png files
