--- README.rst
@@ -1,51 +1,0 @@
-======
-PyCORN
-======
-
-A script to extract data from UNICORN result (.res) files and plot them.
-
-.. image:: https://raw.githubusercontent.com/pyahmed/PyCORN/master/samples/sample1_2009Jun16no001_plot.jpg
-
-A script extract data from .res (results) files generated by UNICORN Chromatography software supplied with AKTA Systems. This script will find all data blocks, extract the data and write out csv-files. If you have matplotlib installed it will also plot all the curves including fractions if present. Plots can be saved in any format supported by matplotlib (default is pdf). Alternatively you may import pycorn into your script and access the data inside res-files.
-
-----
-
-News
-----
-PyCORN-WEB launched:
-PyCORN is now available as a web-service. Just drop your res/zip file, get a preview of the plot (generated with plotly) - adjust as needed and download a high-quality PNG file. No need to open a terminal, no Python installation etc required :-)
-
-visit: http://pycornweb.appspot.com
-
-=======
-
-v0.19 released:
-
-- FIXED: Injection points not recognized properly (see https://github.com/pyahmed/PyCORN/issues/12)
-
-
-Installation
-------------
-- pip install pycorn
-
-
-Known issues
-------------
-- See https://github.com/pyahmed/PyCORN/issues
-
-
-Requirements
-------------
-- Python 2.7 or 3.x (Tested on Linux / Windows 7 / Mac OSX) 
-- optional: matplotlib (for plotting)
-- optional: xlsxwriter (for xlsx-output)
-
-Usage
------
-- See pycorn/docs/USAGE_pycorn-bin.txt if you want to extract/plot data directly
-- See pycorn/docs/USAGE_pycorn_module.txt if you want to import it into your script
-
-
-License
--------
-- GPLv2 see docs/LICENSE.txt
