--- pycorn/docs/CHANGES.txt
@@ -1,58 +1,0 @@
-v0.18
-======
-- FIXED: xmin being ignored when 0
-- FIXED: Files with capitalised name ending are not loaded
-
-v0.17
-======
-- Added experimental support for extracting and plotting UNICORN6 zip bundles
-
-v0.16b
-======
-- Fixed line ending on non-Windows setups
-
-v0.15
-======
-- Added option to disable title for plot
-- Added option to disable plotting inject marks
-- Added option to disable legend
-
-v0.14
-======
-- Inject mark is also plotted
-- xlsx output added (requires xlsxwriter)
-- Fixed par1-error when selecting None
-- Fixed par1/2-error when selecting data blocks with longer data name
-- Minor changes & fixes
-- Included Docs
-
-v0.13
-======
-- Conversion to of PyCORN to class-structure, allows using PyCORN in other scripts (by Wendell Smith)
-- Plots are now UV + user selectable 2nd/3rd data block (2nd y-axis is Conductivity by default)
-- Made some changes to adapt for PyPI
-- Installation should place pycorn-bin.py in Python/Scripts-folder
-
-v0.12
-======
-- Logbook is extracted and written to txt-file
-- Some parts rewritten
-- Fractions are also plotted
-- Improved scaling of y-axis (still not perfect)
-- Option to set start/end-point for plotting
-- Handles files that have been analyzed and re-saved
-
-
-
-v0.11
-======
-- changed data structure (now mostly dicts)
-- made plotting independent of writing csv/txt files
-- all plots contain (if present) fractions
-- x-axis scaling is based on first and last fraction
-- plots can be saved now in all matplotlib-supported formats (default: pdf)
-
-
-v0.1
-======
-- Initial release
